requirements.txt
urllib3==1.26.18
requests>=2.28,<3
PyYAML>=6.0
jsonschema>=4.18
pytest>=7
